Make App a single file component
