--- src/App.global.css
@@ -1,9 +1,0 @@
-a {
-  color:inherit;
-  text-decoration:inherit;
-}
-
-h1 {
-  font-size: 26px;
-  letter-spacing: 0.65px;
-}
