--- src/App.scss
@@ -1,23 +1,0 @@
-#app {
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  display: flex;
-  flex-direction: column;
-
-  &.quick-id-hidden {
-    padding-bottom: 0;
-  }
-
-  .ae-banner {
-    position: fixed;
-    left: 0;
-    right: 0;
-    z-index: auto;
-
-    img {
-      height: 22px;
-      margin-right: 4px;
-      vertical-align: text-bottom;
-    }
-  }
-}
