+++ src/App.vue
@@ -29,6 +29,67 @@
   </ae-main>
 </template>
 
-<style scoped src="./App.scss" lang="scss" />
-<style src="./App.global.css" />
-<script src="./App.js" />
+<script>
+import { mapState } from 'vuex'
+import { AeMain, AeBanner, AeButton } from '@aeternity/aepp-components'
+import FooterMobile from './components/FooterMobile'
+import FooterDesktop from './components/FooterDesktop'
+import RemoveAppModal from './components/RemoveAppModal.vue'
+import AlertModal from './components/AlertModal'
+
+export default {
+  components: {
+    AeMain,
+    AeBanner,
+    AeButton,
+    RemoveAppModal,
+    AlertModal,
+    AppFooter: process.env.IS_MOBILE_DEVICE ? FooterMobile : FooterDesktop
+  },
+  computed: mapState(['notification']),
+  created: function () {
+    // set domain to base host because of iframe cross domain policy, very nice hardcoded urls
+    if (document.domain.includes('aepps.com')) {
+      document.domain = 'aepps.com'
+    } else if (document.domain.includes('aepps.dev')) {
+      document.domain = 'aepps.dev'
+    }
+  }
+}
+</script>
+
+<style scoped lang="scss">
+#app {
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  display: flex;
+  flex-direction: column;
+
+  &.quick-id-hidden {
+    padding-bottom: 0;
+  }
+
+  .ae-banner {
+    position: fixed;
+    left: 0;
+    right: 0;
+    z-index: auto;
+
+    img {
+      height: 22px;
+      margin-right: 4px;
+      vertical-align: text-bottom;
+    }
+  }
+
+  /deep/ a {
+    color: inherit;
+    text-decoration: inherit;
+  }
+
+  /deep/ h1 {
+    font-size: 26px;
+    letter-spacing: 0.65px;
+  }
+}
+</style>
